pyproject.toml
[project]
name = "cpex"
version = "0.1.0.dev4"
description = "CPEX - ContextForge Plugin Extensibility Framework"
classifiers = [
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Software Development :: Libraries :: Application Frameworks",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
license-files = ["LICENSE*"]
readme = "README.md"
authors = [
    { name = "Fred Araujo", email = "frederico.araujo@ibm.com" },
    { name = "Mihai Criveti", email = "redacted@ibm.com"},
    { name = "Teryl Taylor", email = "terylt@ibm.com" },
]
maintainers = [
    { name = "Fred Araujo", email = "frederico.araujo@ibm.com" },
    { name = "Jonathan Springer", email = "redacted@ibm.com" },
    { name = "Teryl Taylor", email = "terylt@ibm.com" },
]
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.133.1",
    "httpx>=0.28.1",
    "httpx[http2]>=0.28.1",
    "jinja2>=3.1.6",
    "mcp>=1.26.0",
    "orjson>=3.11.7",
    "prometheus-fastapi-instrumentator>=7.1.0",
    "prometheus_client>=0.24.1",
    "pydantic-settings>=2.13.1",
    "pydantic>=2.12.5",
    "pyyaml>=6.0.3",
]

[project.scripts]
cpex = "cpex.tools.cli:main"

[project.optional-dependencies]

# dev dependencies
dev = [
    "bandit>=1.8.6",
    "interrogate>=1.7.0",
    "mypy>=1.18.2",
    "pyroma>=5.0",
    "pytest>=8.4.2",
    "pytest-asyncio>=1.2.0",
    "pytest-cov>=7.0.0",
    "pytest-env>=1.1.5",
    "pytest-xdist>=3.8.0",
    "radon>=6.0.1",
    "ruff>=0.13.3",
    "twine>=6.2.0",
    "types-PyYAML>=6.0.0",
    "vulture>=2.14",
]

# documentation support
docs = [
    "griffe-pydantic",
    "griffe_inherited_docstrings",
    "mkdocs-awesome-nav",
    "mkdocs-click",
    "mkdocs-enumerate-headings-plugin",
    "mkdocs-include-markdown-plugin>=7.1.8",
    "mkdocs-jupyter",
    "mkdocs-material",
    "mkdocstrings[python]",
]

# gRPC transport support (higher performance than MCP/HTTP)
grpc = [
    "grpcio>=1.70.0",
    "grpcio-tools>=1.70.0",
    "protobuf>=5.29.0",
]

# CLI dependencies
cli = [
    "chardet>=3.0.2,<6",  # pin to range supported by requests (binaryornot pulls chardet 7+)
    "cookiecutter>=2.6.0",
    "typer>=0.21.2", # docling requires typer < 0.22
]

# All extras
all = ["cpex[grpc,cli]"]

[tool.interrogate]
exclude = ["cpex/templates", "*_pb2.py", "*_pb2_grpc.py"]
fail-under = 100

[tool.radon]
exclude = "cpex/templates/*,*_pb2.py,*_pb2_grpc.py"

[tool.vulture]
min_confidence = 80
exclude = ["cpex/templates", "cpex/framework/observability.py", "*_pb2.py", "*_pb2_grpc.py"]

[tool.ruff]
line-length = 120
exclude = [
    "cpex/templates",
]

# Linter configuration
[tool.ruff.lint]
# Enable Pyflakes (`F`) and a subset of the pycodestyle (`E`)  codes by default.
# Also "D1" for docstring present checks.
# "PL" enables Pylint rules natively in ruff (replaces pylint for these checks).
select = ["E3", "E4", "E7", "E9", "F", "D1", "PL", "I"]
ignore = [
    # ---- Rules already disabled in .pylintrc (intentionally relaxed) ----
    "PLR0904",  # too-many-public-methods  (pylint R0904)
    "PLR0912",  # too-many-branches        (pylint R0912)
    "PLR0913",  # too-many-arguments       (pylint R0913)
    "PLR0914",  # too-many-locals          (pylint R0914)
    "PLR0915",  # too-many-statements      (pylint R0915)
    # ---- Accepted codebase patterns (high violation count, tolerated by pylint scoring) ----
    "PLC0415",  # import-outside-top-level (pylint C0415) — lazy-import patterns
    "PLW0603",  # global-statement         (pylint W0603) — config/singleton patterns
    "PLR0911",  # too-many-returns
    "PLR0917",  # too-many-positional-args
    "PLR1702",  # too-many-nested-blocks
    # ---- Ruff-only PL rules with no pylint equivalent (out of migration scope) ----
    "PLC1901",  # compare-to-empty-string  (deprecated in pylint)
    "PLC2701",  # import-private-name      (ruff-only)
    "PLR2004",  # magic-value-comparison   (ruff-only)
    "PLR5501",  # collapsible-else-if      (ruff-only)
    "PLR6104",  # non-augmented-assignment (ruff-only)
    "PLR6201",  # literal-membership       (ruff-only)
    "PLR6301",  # no-self-use              (ruff-only)
    "PLW2901",  # redefined-loop-name      (ruff-only)
]

preview = true

# Allow fix for all enabled rules (when `--fix`) is provided.
fixable = ["ALL"]
unfixable = []

# Ignore D1 (docstring checks) and Pylint checks in tests and other non-production code
[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = ["D1", "PL"]
"scripts/**/*.py" = ["D1", "PL"]
".github/**/*.py" = ["D1", "PL"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["cpex"]
